pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sas-to-snowflake-migrator"
version = "0.1.0"
description = "SAS to Snowflake/Databricks Migration Toolkit"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "pyreadstat>=1.2.0",
    "sas7bdat>=2.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
]

[project.scripts]
sas-migrator = "src.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
